+++ packages/sit-onyx/src/components/OnyxForm/types.ts
@@ -1,4 +1,4 @@
 import type { DensityProp } from "../../composables/density";
-import type { FormInjectedProps } from "./OnyxForm";
+import type { FormInjectedProps } from "./OnyxForm.core";
 
 export type OnyxFormProps = Partial<FormInjectedProps> & DensityProp;
