+++ packages/sit-onyx/src/components/OnyxIconButton/types.ts
@@ -1,7 +1,7 @@
 import type { DensityProp } from "../../composables/density";
 import type { AutofocusProp } from "../../types";
 import type { ButtonColor, ButtonType } from "../OnyxButton/types";
-import type { FormInjected } from "../OnyxForm/OnyxForm";
+import type { FormInjected } from "../OnyxForm/OnyxForm.core";
 
 export type OnyxIconButtonProps = DensityProp &
   AutofocusProp & {
