+++ packages/sit-onyx/src/components/OnyxRadioGroup/OnyxRadioGroup.vue
@@ -6,7 +6,7 @@ import type { SelectOptionValue } from "../../types";
 import OnyxHeadline from "../OnyxHeadline/OnyxHeadline.vue";
 import OnyxRadioButton from "../OnyxRadioButton/OnyxRadioButton.vue";
 import type { OnyxRadioGroupProps } from "./types";
-import { FORM_INJECTED_SYMBOL, useFormContext } from "../OnyxForm/OnyxForm";
+import { FORM_INJECTED_SYMBOL, useFormContext } from "../OnyxForm/OnyxForm.core";
 
 const props = withDefaults(defineProps<OnyxRadioGroupProps<TValue>>(), {
   name: () => useId() ?? "", // the name must be globally unique
