+++ packages/sit-onyx/src/components/OnyxSwitch/types.ts
@@ -2,7 +2,7 @@ import type { DensityProp } from "../../composables/density";
 import type { RequiredMarkerProp } from "../../composables/required";
 import type { CustomValidityProp } from "../../composables/useCustomValidity";
 import type { AutofocusProp, TruncationType } from "../../types";
-import type { FormInjected } from "../OnyxForm/OnyxForm";
+import type { FormInjected } from "../OnyxForm/OnyxForm.core";
 
 export type OnyxSwitchProps = DensityProp &
   RequiredMarkerProp &
