+++ packages/sit-onyx/src/components/OnyxTextarea/OnyxTextarea.vue
@@ -5,7 +5,7 @@ import { useCustomValidity } from "../../composables/useCustomValidity";
 import OnyxFormElement from "../OnyxFormElement/OnyxFormElement.vue";
 import OnyxSkeleton from "../OnyxSkeleton/OnyxSkeleton.vue";
 import type { OnyxTextareaProps } from "./types";
-import { FORM_INJECTED_SYMBOL, useFormContext } from "../OnyxForm/OnyxForm";
+import { FORM_INJECTED_SYMBOL, useFormContext } from "../OnyxForm/OnyxForm.core";
 
 const props = withDefaults(defineProps<OnyxTextareaProps>(), {
   modelValue: "",
